Changed remaining demand interfaces and classes to use single-precision
This commit gets rid of the lingering instances of double precision in
demand structures and their relevant methods. This method also removes
the AreaSpecificTripGenerator classes and the threads associated with
the former enum-based TripPurpose structure. EnhancedGraphs are also
read in parallel now (improvements still can be made in the concurrency
structure here)

+++ src/edu/utexas/wrap/demand/containers/AutoFixedSizeMatrix.java
@@ -26,15 +26,15 @@ public class AutoFixedSizeMatrix implements ODMatrix {
 	}
 
 	@Override
-	public Float getDemand(TravelSurveyZone origin, TravelSurveyZone destination) {
+	public float getDemand(TravelSurveyZone origin, TravelSurveyZone destination) {
 		DemandMap orig = origins[origin.getOrder()];
-		return orig == null? 0.0f : orig.get(destination).floatValue();
+		return orig == null? 0.0f : orig.get(destination);
 	}
 
 	@Override
 	public void put(TravelSurveyZone origin, TravelSurveyZone destination, Float demand) {
 		if (origins[origin.getOrder()] == null) origins[origin.getOrder()] = new FixedSizeDemandMap(graph);
-		origins[origin.getOrder()].put(destination, demand.doubleValue());
+		origins[origin.getOrder()].put(destination, demand);
 	}
 
 	@Override
