+++ src/edu/utexas/wrap/demand/containers/FixedSizePAMap.java
@@ -35,13 +35,13 @@ public class FixedSizePAMap implements PAMap {
 	}
 
 	@Override
-	public Float getAttractions(TravelSurveyZone z) {
-		return attrs.get(z).floatValue();
+	public float getAttractions(TravelSurveyZone z) {
+		return attrs.get(z);
 	}
 
 	@Override
-	public Float getProductions(TravelSurveyZone z) {
-		return prods.get(z).floatValue();
+	public float getProductions(TravelSurveyZone z) {
+		return prods.get(z);
 	}
 
 	@Override
@@ -51,12 +51,12 @@ public class FixedSizePAMap implements PAMap {
 
 	@Override
 	public void putAttractions(TravelSurveyZone z, Float amt) {
-		attrs.put(z, amt.doubleValue());
+		attrs.put(z, amt);
 	}
 
 	@Override
 	public void putProductions(TravelSurveyZone z, Float amt) {
-		prods.put(z, amt.doubleValue());
+		prods.put(z, amt);
 	}
 
 	@Override
